+++ engine/src/main/java/io/zeebe/engine/processor/workflow/handlers/catchevent/StartEventEventOccurredHandler.java
@@ -16,6 +16,7 @@ import io.zeebe.engine.state.ZeebeState;
 import io.zeebe.engine.state.deployment.DeployedWorkflow;
 import io.zeebe.engine.state.deployment.WorkflowState;
 import io.zeebe.engine.state.instance.EventTrigger;
+import io.zeebe.engine.state.message.MessageState;
 import io.zeebe.protocol.impl.record.value.workflowinstance.WorkflowInstanceRecord;
 import io.zeebe.protocol.record.intent.WorkflowInstanceIntent;
 import io.zeebe.protocol.record.value.BpmnElementType;
@@ -28,20 +29,23 @@ public class StartEventEventOccurredHandler<T extends ExecutableCatchEventElemen
 
   private final WorkflowInstanceRecord record = new WorkflowInstanceRecord();
   private final WorkflowState state;
+  private final MessageState messageState;
   private final KeyGenerator keyGenerator;
 
-  public StartEventEventOccurredHandler(ZeebeState zeebeState) {
+  public StartEventEventOccurredHandler(final ZeebeState zeebeState) {
     this(null, zeebeState);
   }
 
-  public StartEventEventOccurredHandler(WorkflowInstanceIntent nextState, ZeebeState zeebeState) {
+  public StartEventEventOccurredHandler(
+      final WorkflowInstanceIntent nextState, final ZeebeState zeebeState) {
     super(nextState);
     this.state = zeebeState.getWorkflowState();
+    this.messageState = zeebeState.getMessageState();
     keyGenerator = zeebeState.getKeyGenerator();
   }
 
   @Override
-  protected boolean handleState(BpmnStepContext<T> context) {
+  protected boolean handleState(final BpmnStepContext<T> context) {
     final WorkflowInstanceRecord event = context.getValue();
     final long workflowKey = event.getWorkflowKey();
     final DeployedWorkflow workflow = state.getWorkflowByKey(workflowKey);
@@ -70,11 +74,20 @@ public class StartEventEventOccurredHandler<T extends ExecutableCatchEventElemen
             .setFlowScopeKey(workflowInstanceKey);
 
     deferEvent(context, workflowKey, workflowInstanceKey, record, triggeredEvent);
+
+    // should mark particular message as already correlated for this instance
+    if (context.getElement().isMessage()) {
+      final long messageKey = triggeredEvent.getEventKey();
+      messageState.putMessageCorrelation(messageKey, workflowInstanceKey);
+    }
+
     return true;
   }
 
   private void createWorkflowInstance(
-      BpmnStepContext<T> context, DeployedWorkflow workflow, long workflowInstanceKey) {
+      final BpmnStepContext<T> context,
+      final DeployedWorkflow workflow,
+      final long workflowInstanceKey) {
     record
         .setBpmnProcessId(workflow.getBpmnProcessId())
         .setWorkflowKey(workflow.getKey())
@@ -88,15 +101,5 @@ public class StartEventEventOccurredHandler<T extends ExecutableCatchEventElemen
             WorkflowInstanceIntent.ELEMENT_ACTIVATING,
             record,
             workflow.getWorkflow());
-  }
-
-  private void deferStartEventRecord(
-      BpmnStepContext<T> context, long workflowInstanceKey, WorkflowInstanceRecord event) {
-    event.setWorkflowInstanceKey(workflowInstanceKey);
-    event.setFlowScopeKey(workflowInstanceKey);
-
-    context
-        .getOutput()
-        .deferRecord(workflowInstanceKey, event, WorkflowInstanceIntent.ELEMENT_ACTIVATING);
   }
 }