+++ exporters/elasticsearch-exporter/src/test/java/io/zeebe/exporter/ElasticsearchExporterFaultToleranceIT.java
@@ -10,6 +10,7 @@ package io.zeebe.exporter;
 import io.zeebe.protocol.record.Record;
 import io.zeebe.test.util.TestUtil;
 import io.zeebe.test.util.record.RecordingExporter;
+import io.zeebe.test.util.socket.SocketUtil;
 import org.elasticsearch.ElasticsearchStatusException;
 import org.junit.Test;
 
@@ -19,6 +20,7 @@ public class ElasticsearchExporterFaultToleranceIT
   @Test
   public void shouldExportEvenIfElasticNotInitiallyReachable() {
     // given
+    elastic.withPort(SocketUtil.getNextAddress().getPort());
     configuration = getDefaultConfiguration();
     configuration.index.prefix = "zeebe";
     esClient = createElasticsearchClient(configuration);