+++ test/src/main/java/io/zeebe/test/EmbeddedBrokerRule.java
@@ -23,8 +23,8 @@ import io.zeebe.servicecontainer.ServiceStartContext;
 import io.zeebe.servicecontainer.ServiceStopContext;
 import io.zeebe.test.util.record.RecordingExporter;
 import io.zeebe.test.util.record.RecordingExporterTestWatcher;
+import io.zeebe.test.util.socket.SocketUtil;
 import io.zeebe.transport.SocketAddress;
-import io.zeebe.transport.impl.util.SocketUtil;
 import io.zeebe.util.FileUtil;
 import io.zeebe.util.TomlConfigurationReader;
 import io.zeebe.util.ZbLogger;
@@ -119,10 +119,10 @@ public class EmbeddedBrokerRule extends ExternalResource {
 
   public static void assignSocketAddresses(final BrokerCfg brokerCfg) {
     final NetworkCfg network = brokerCfg.getNetwork();
-    brokerCfg.getGateway().getNetwork().setPort(SocketUtil.getNextAddress().port());
-    network.getCommandApi().setPort(SocketUtil.getNextAddress().port());
-    network.getInternalApi().setPort(SocketUtil.getNextAddress().port());
-    network.getMonitoringApi().setPort(SocketUtil.getNextAddress().port());
+    brokerCfg.getGateway().getNetwork().setPort(SocketUtil.getNextAddress().getPort());
+    network.getCommandApi().setPort(SocketUtil.getNextAddress().getPort());
+    network.getInternalApi().setPort(SocketUtil.getNextAddress().getPort());
+    network.getMonitoringApi().setPort(SocketUtil.getNextAddress().getPort());
   }
 
   @Override